+++ .husky/commit-msg
@@ -2,3 +2,4 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 npx --no-install commitlint --edit
+npx --no-install commitlint --edit
