+++ android/app/src/main/java/com/ctask/mobile/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.boilerplate.typescript
+package com.ctask.mobile
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
@@ -11,7 +11,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "RNTypescriptBoilerplate"
+  override fun getMainComponentName(): String = "ctask"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
