+++ ios/Podfile
@@ -25,7 +25,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'RNTypescriptBoilerplate' do
+target 'CTask' do
   config = use_native_modules!
 
   use_react_native!(
@@ -39,7 +39,7 @@ target 'RNTypescriptBoilerplate' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'RNTypescriptBoilerplateTests' do
+  target 'CTaskTests' do
     inherit! :complete
     # Pods for testing
   end
