+++ ios/ctask.xcodeproj/xcshareddata/xcschemes/ctask.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "RNTypescriptBoilerplate.app"
-               BlueprintName = "RNTypescriptBoilerplate"
-               ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+               BuildableName = "CTask.app"
+               BlueprintName = "CTask"
+               ReferencedContainer = "container:ctask.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RNTypescriptBoilerplateTests.xctest"
-               BlueprintName = "RNTypescriptBoilerplateTests"
-               ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+               BuildableName = "CTaskTests.xctest"
+               BlueprintName = "CTaskTests"
+               ReferencedContainer = "container:ctask.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTypescriptBoilerplate.app"
-            BlueprintName = "RNTypescriptBoilerplate"
-            ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+            BuildableName = "CTask.app"
+            BlueprintName = "CTask"
+            ReferencedContainer = "container:ctask.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTypescriptBoilerplate.app"
-            BlueprintName = "RNTypescriptBoilerplate"
-            ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+            BuildableName = "CTask.app"
+            BlueprintName = "CTask"
+            ReferencedContainer = "container:ctask.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
