+++ ios/ctask.xcworkspace/contents.xcworkspacedata
@@ -2,12 +2,9 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RNTypescriptBoilerplate.xcodeproj">
-   </FileRef>
-   <FileRef
       location = "group:Pods/Pods.xcodeproj">
    </FileRef>
    <FileRef
-      location = "group:rntypescriptBoilerplate.xcodeproj">
+      location = "group:ctask.xcodeproj">
    </FileRef>
 </Workspace>
