+++ ios/ctask/Images.xcassets/AppIcon.appiconset/Contents.json
