+++ ios/ctaskTests/Info.plist
