+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-native-typescript-boilerplate",
+  "name": "ctask",
   "version": "4.0.2",
   "private": false,
   "scripts": {
